Update to Crosswalk embedding API v2.
Embedding API v2 is in Crosswalk-8.
See https://crosswalk-project.org/jira/browse/XWALK-2002 for details.
